Split track tools to a separate file.
