+++ src/animation/offline/tools/import2ozz.cc
@@ -162,6 +162,17 @@ int OzzImporter::operator()(int _argc, const char** _argv) {
 
   return EXIT_SUCCESS;
 }
+
+ozz::String::Std OzzImporter::BuildFilename(const char* _filename,
+                                            const char* _data_name) const {
+  ozz::String::Std output(_filename);
+
+  for (size_t asterisk = output.find('*'); asterisk != std::string::npos;
+       asterisk = output.find('*')) {
+    output.replace(asterisk, 1, _data_name);
+  }
+  return output;
+}
 }  // namespace offline
 }  // namespace animation
 }  // namespace ozz
